backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
pyaudio==0.2.13
requests==2.31.0
numpy==1.24.3
